reimplement recursion for indentation depth
It now does the following:

If we first see an opening keyword before our current line, then we need
to increase our indentation relative to that line.
Otherwise if we first see a non-empty line, then we should indent our
current line to the same depth.
If none of the above apply, then do not indent at all.